requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
docopt==0.6.2
idna==3.6
mysqlclient==2.2.1
pip==22.3.1
pipreqs==0.4.13
pytz==2023.3.post1
requests==2.31.0
rest-framework-simplejwt==0.0.2
setuptools==65.5.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wheel==0.38.4
yarg==0.1.9
